Move geojson creation from base class to tile map service class
